#430: a functor that narrows its member re-introduces the wrap the widening fixed
CvCity.h declared the getCultureTimes100 map-functor's return as `int` while the
member it wraps returns int64_t. Culture is precisely the value that outgrew 32
bits -- it accumulates the x100 rate every turn and never decays -- so the functor
narrowed it straight back on every call.

Two things settle that this was an oversight rather than a choice: its immediate
neighbours calcCorporateMaintenance and getRealPopulation are ALREADY declared
int64_t, so this one was simply missed when culture was widened; and nothing uses
the functor today, so it was latent rather than live -- the first algorithm to
reach for it would have silently truncated.

It was also 50 of the build's 90 C4244 emissions on its own: an inline in a header
re-emits once per translation unit. Assert goes 90 -> 40, and every remaining
warning is now a unique site rather than an echo.

+++ Sources/Engine/CvCity.h
@@ -1970,7 +1970,9 @@ public:
 		DECLARE_MAP_FUNCTOR_CONST_1(CvCity, bool, isCoastal, int);
 		DECLARE_MAP_FUNCTOR_CONST_1(CvCity, bool, hasBuilding, BuildingTypes);
 		DECLARE_MAP_FUNCTOR_CONST_1(CvCity, bool, isActiveBuilding, BuildingTypes);
-		DECLARE_MAP_FUNCTOR_CONST_1(CvCity, int, getCultureTimes100, PlayerTypes);
+		//	int64_t like the member it wraps: culture accumulates every turn and never decays, so it outgrew `int`
+		//	-- declaring the functor `int` narrowed it back on every call, which is the wrap the widening fixed.
+		DECLARE_MAP_FUNCTOR_CONST_1(CvCity, int64_t, getCultureTimes100, PlayerTypes);
 		DECLARE_MAP_FUNCTOR_CONST_1(CvCity, const CvPlotGroup*, plotGroup, PlayerTypes);
 
 		DECLARE_MAP_FUNCTOR_CONST_2(CvCity, bool, isRevealed, TeamTypes, bool);
